requirements.txt
streamlit
pandas
altair
requests
psycopg2-binary
playwright
